Change logic for delete `employee`

+++ src/main/java/io/github/terminalemployeemanager/service/EmployeeService.java
@@ -34,7 +34,12 @@ public class EmployeeService {
         employeeRepository.save(employee);
     }
     public void deleteEmployee(Long id){
-        employeeRepository.deleteById(id);
+        if (employeeRepository.existsById(id)) {
+            employeeRepository.deleteById(id);
+            System.out.println("Employee with " + id + " successfully deleted.");
+        } else {
+            System.out.println("Employee not found");
+        }
     }
